requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
alembic==1.14.1
asyncpg==0.31.0
attrs==26.1.0
bcrypt==3.2.2
cachetools==6.2.6
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cohere==5.21.1
colorama==0.4.6
cryptography==47.0.0
dataclasses-json==0.6.7
deprecation==2.1.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.136.1
fastavro==1.12.2
filelock==3.29.0
frozenlist==1.8.0
fsspec==2026.3.0
greenlet==3.4.0
grpcio==1.80.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.3
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.12.0
hyperframe==6.1.0
idna==3.13
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-classic==1.0.4
langchain-cohere==0.5.1
langchain-community==0.4.1
langchain-core==1.3.2
langchain-protocol==0.0.12
langchain-text-splitters==1.1.2
langgraph==1.1.9
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.11
langgraph-sdk==0.3.13
langsmith==0.7.37
markdown-it-py==4.0.0
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.4
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.2
passlib==1.7.4
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
portalocker==3.2.0
postgrest==2.29.0
propcache==0.4.1
protobuf==7.34.1
psycopg2==2.9.12
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.14.0
pydantic_core==2.41.5
Pygments==2.20.0
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pypdfium2==5.7.1
pyroaring==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
pywin32==311
PyYAML==6.0.3
qdrant-client==1.17.1
realtime==2.29.0
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.4
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
storage3==2.29.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.29.0
supabase-auth==2.29.0
supabase-functions==2.29.0
tenacity==9.1.4
tokenizers==0.22.2
tqdm==4.67.3
typer==0.25.0
types-PyYAML==6.0.12.20260408
types-requests==2.33.0.20260408
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.46.0
websockets==15.0.1
xxhash==3.7.0
yarl==1.23.0
zstandard==0.25.0
